Move DnaModifiers
