+++ crates/holochain_integrity_types/src/prelude.rs
@@ -5,7 +5,7 @@ pub use crate::action::*;
 pub use crate::capability::*;
 pub use crate::chain::*;
 pub use crate::countersigning::*;
-pub use crate::dna_properties::*;
+pub use crate::dna_modifiers::*;
 pub use crate::entry::*;
 pub use crate::entry_def::*;
 pub use crate::genesis::*;
